--- requirements.txt
@@ -1,1 +1,0 @@
-progressbar==2.5
