+++ src/mc_server_python/__main__.py
@@ -1,0 +1,9 @@
+from mc_server_python.main import main
+
+
+def cli():
+    main()
+
+
+if __name__ == '__main__':
+    cli()
